+++ particula/next/gas/species.py
@@ -76,128 +76,128 @@ class GasSpecies:
         return np.float64(self.mass) if self.condensable else np.float64(0)
 
 
-@dataclass
-class Gas:
-    """
-    Represents a mixture of gas species, including properties such as
-    temperature, total pressure, and a list of gas species components.
+# @dataclass
+# class Gas:
+#     """
+#     Represents a mixture of gas species, including properties such as
+#     temperature, total pressure, and a list of gas species components.
 
-    Attributes:
-    - temperature (float): The temperature of the gas mixture.
-    - total_pressure (float): The total pressure of the gas mixture.
-    - components (List[GasSpecies]): A list of GasSpecies objects
-        representing the components of the gas mixture.
+#     Attributes:
+#     - temperature (float): The temperature of the gas mixture.
+#     - total_pressure (float): The total pressure of the gas mixture.
+#     - components (List[GasSpecies]): A list of GasSpecies objects
+#         representing the components of the gas mixture.
 
-    Methods:
-    - add_species: Adds a gas species to the mixture.
-    - remove_species: Removes a gas species from the mixture by name.
-    - get_mass: Returns the mass of a specified species or the masses of all
-        species in the gas mixture as an np.ndarray.
-    - get_mass_condensable: Returns the mass of a specific condensable species
-        or the masses of all condensable species in the gas mixture as an
-        np.ndarray.
+#     Methods:
+#     - add_species: Adds a gas species to the mixture.
+#     - remove_species: Removes a gas species from the mixture by name.
+#     - get_mass: Returns the mass of a specified species or the masses of all
+#         species in the gas mixture as an np.ndarray.
+#     - get_mass_condensable: Returns the mass of a specific condensable species
+#         or the masses of all condensable species in the gas mixture as an
+#         np.ndarray.
 
-    """
-    temperature: float = 298.15
-    total_pressure: float = 101325
-    components: list[GasSpecies] = field(default_factory=list)
+#     """
+#     temperature: float = 298.15
+#     total_pressure: float = 101325
+#     components: list[GasSpecies] = field(default_factory=list)
 
-    def add_species(
-            self,
-            name: str,
-            mass: float,
-            vapor_pressure: Optional[float] = None,
-            condensable: bool = False) -> None:
-        """
-        Adds a gas species to the mixture.
+#     def add_species(
+#             self,
+#             name: str,
+#             mass: float,
+#             vapor_pressure: Optional[float] = None,
+#             condensable: bool = False) -> None:
+#         """
+#         Adds a gas species to the mixture.
 
-        Parameters:
-        - name (str): The name of the gas species.
-        - mass (float): The mass of the gas species.
-        - vapor_pressure (Optional[float]): The vapor pressure of the gas
-            species. None if not applicable.
-        - condensable (bool): Indicates whether the gas species is
-            condensable.
-        """
-        species = GasSpecies(name, mass, vapor_pressure, condensable)
-        self.components.append(species)
+#         Parameters:
+#         - name (str): The name of the gas species.
+#         - mass (float): The mass of the gas species.
+#         - vapor_pressure (Optional[float]): The vapor pressure of the gas
+#             species. None if not applicable.
+#         - condensable (bool): Indicates whether the gas species is
+#             condensable.
+#         """
+#         species = GasSpecies(name, mass, vapor_pressure, condensable)
+#         self.components.append(species)
 
-    def remove_species(self, name: str) -> None:
-        """
-        Removes a gas species from the mixture by name.
+#     def remove_species(self, name: str) -> None:
+#         """
+#         Removes a gas species from the mixture by name.
 
-        Parameters:
-        - name (str): The name of the gas species to be removed.
-        """
-        self.components = [c for c in self.components if c.name != name]
+#         Parameters:
+#         - name (str): The name of the gas species to be removed.
+#         """
+#         self.components = [c for c in self.components if c.name != name]
 
-    def get_mass(self, name: Optional[str] = None) -> NDArray[np.float64]:
-        """
-        Returns the mass of a specified species or the masses of all species
-        in the gas mixture as an np.ndarray.
+#     def get_mass(self, name: Optional[str] = None) -> NDArray[np.float64]:
+#         """
+#         Returns the mass of a specified species or the masses of all species
+#         in the gas mixture as an np.ndarray.
 
-        If a name is specified, the method returns an array containing the
-        mass of the named species. If the specified name is not found in the
-        mixture, a ValueError is raised. If no name is specified, it returns
-        an array of the masses of all species in the mixture.
+#         If a name is specified, the method returns an array containing the
+#         mass of the named species. If the specified name is not found in the
+#         mixture, a ValueError is raised. If no name is specified, it returns
+#         an array of the masses of all species in the mixture.
 
-        Parameters:
-            name (Optional[str]): The name of the gas species for which to
-            return the mass. If None, returns masses of all species.
+#         Parameters:
+#             name (Optional[str]): The name of the gas species for which to
+#             return the mass. If None, returns masses of all species.
 
-        Returns:
-            NDArray[np.float64]: An array containing the requested mass(es).
+#         Returns:
+#             NDArray[np.float64]: An array containing the requested mass(es).
 
-        Raises:
-            ValueError: If the specified name is not found in the mixture.
-        """
-        if name:
-            matching_masses = [component.get_mass()
-                               for component in self.components
-                               if component.name == name]
-            if not matching_masses:
-                raise ValueError(
-                    f"Gas species '{name}' not found in the mixture.")
-            return np.array(matching_masses, dtype=np.float64)
+#         Raises:
+#             ValueError: If the specified name is not found in the mixture.
+#         """
+#         if name:
+#             matching_masses = [component.get_mass()
+#                                for component in self.components
+#                                if component.name == name]
+#             if not matching_masses:
+#                 raise ValueError(
+#                     f"Gas species '{name}' not found in the mixture.")
+#             return np.array(matching_masses, dtype=np.float64)
 
-        # Return the masses of all components if no name is specified.
-        masses = [component.get_mass() for component in self.components]
-        return np.array(masses, dtype=np.float64)
+#         # Return the masses of all components if no name is specified.
+#         masses = [component.get_mass() for component in self.components]
+#         return np.array(masses, dtype=np.float64)
 
-    def get_mass_condensable(
-            self, name: Optional[str] = None) -> NDArray[np.float64]:
-        """
-        Returns the mass of a specific condensable species or the masses of
-        all condensable species in the gas mixture as an np.ndarray. If a name
-        is provided, only the mass of that specific condensable species is
-        returned.
+#     def get_mass_condensable(
+#             self, name: Optional[str] = None) -> NDArray[np.float64]:
+#         """
+#         Returns the mass of a specific condensable species or the masses of
+#         all condensable species in the gas mixture as an np.ndarray. If a name
+#         is provided, only the mass of that specific condensable species is
+#         returned.
 
-        Parameters:
-            name (Optional[str]): The name of the specific condensable gas
-            species to retrieve the mass for. If None (the default), the
-            masses of all condensable species are returned.
+#         Parameters:
+#             name (Optional[str]): The name of the specific condensable gas
+#             species to retrieve the mass for. If None (the default), the
+#             masses of all condensable species are returned.
 
-        Returns:
-            NDArray[np.float64]: An array of the mass of the specified
-            condensable species, or an array of the masses of all condensable
-            species in the gas mixture.
+#         Returns:
+#             NDArray[np.float64]: An array of the mass of the specified
+#             condensable species, or an array of the masses of all condensable
+#             species in the gas mixture.
 
-        Raises:
-            ValueError: If a specific species name is provided but not found
-            in the mixture, or if it's not condensable.
-        """
-        if name:
-            for component in self.components:
-                if component.name == name:
-                    if component.is_condensable():
-                        return np.array(
-                            [component.get_mass()], dtype=np.float64)
-                    raise ValueError(
-                        f"Gas species '{name}' is not condensable.")
-            raise ValueError(f"Gas species '{name}' not found in the mixture.")
+#         Raises:
+#             ValueError: If a specific species name is provided but not found
+#             in the mixture, or if it's not condensable.
+#         """
+#         if name:
+#             for component in self.components:
+#                 if component.name == name:
+#                     if component.is_condensable():
+#                         return np.array(
+#                             [component.get_mass()], dtype=np.float64)
+#                     raise ValueError(
+#                         f"Gas species '{name}' is not condensable.")
+#             raise ValueError(f"Gas species '{name}' not found in the mixture.")
 
-        masses_condensable = [
-            component.get_mass()
-            for component in self.components if component.is_condensable()
-        ]
-        return np.array(masses_condensable, dtype=np.float64)
+#         masses_condensable = [
+#             component.get_mass()
+#             for component in self.components if component.is_condensable()
+#         ]
+#         return np.array(masses_condensable, dtype=np.float64)
