+++ particula/next/gas/tests/gas_collection_test.py
@@ -5,8 +5,8 @@
 import pytest
 import numpy as np
 from particula.next.gas import Gas, GasBuilder
-from particula.next.gas_species import GasSpeciesBuilder
-from particula.next.gas_vapor_pressure import ConstantVaporPressureStrategy
+from particula.next.gas.gas_species import GasSpeciesBuilder
+from particula.next.gas.gas_vapor_pressure import ConstantVaporPressureStrategy
 
 
 def test_gas_initialization():
