+++ particula/next/tests/aerosol_test.py
@@ -9,8 +9,8 @@ import pytest
 import numpy as np
 from particula.next.aerosol import Aerosol
 from particula.next.gas import GasBuilder
-from particula.next.gas_species import GasSpeciesBuilder, GasSpecies
-from particula.next.gas_vapor_pressure import ConstantVaporPressureStrategy
+from particula.next.gas.gas_species import GasSpeciesBuilder, GasSpecies
+from particula.next.gas.gas_vapor_pressure import ConstantVaporPressureStrategy
 from particula.next.particle import Particle, particle_strategy_factory
 from particula.next.particle_activity import MassIdealActivity
 from particula.next.surface import surface_strategy_factory
